+++ src/main/kotlin/com/loadingbyte/cinecred/ui/styling/StyleForm.kt
@@ -266,7 +266,7 @@ class StyleForm<S : Style>(
                 toString = Locale::getDisplayName, widthSpec
             )
             Color::class.java -> ColorWellWidget(allowAlpha = colorConstr?.allowAlpha ?: true, widthSpec)
-            Resolution::class.java -> ResolutionWidget(defaultCustom = PRESET_GLOBAL.resolution)
+            Resolution::class.java -> ResolutionWidget()
             FPS::class.java -> FPSWidget(widthSpec)
             FontFeature::class.java -> FontFeatureWidget()
             else -> when {